Unique headstone UV fix
This fix allows for all unique headstones to have a texture and it has a fix to prevent the normals of the house from causes visual glitches.
The Dade sign is also fixed to have text on the front and the back
A texture is also added to the pyramids.
PA-0-2 might not present a texture.